[cpp] Add validating storage.
This patch adds a ValidatingStorage object, which wraps any implementation
of Storage by adding checksum and timestamp when saving data. When reading
data, the object verifies the timestamp freshness and checksum validity.
If these checks fail, then ValidatingStorage behaves as if there's no data
in storage. This behavior would force a Retriever object to re-download
validation rules from the server.

A data timestamp is fresh if it's within a month of the current time.

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@174 38ededc0-08b8-5190-f2ac-b31f878777ad