Update re2.gyp to work when building outside of Chromium.
The variable "component" is not going to be defined when building
libaddressinput following the instructions in README, so it's not
permissible to test the value of this variable.

Instead, add a requirement for at least version 20140111+dfsg-1 of the
RE2 package (which will be needed for the set_never_capture() option, to
be used by rule.cc, anyway), and then the package also provides a shared
library so that it becomes possible to specify linking with libre2 using
a normal "-lre2" linker directive.

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@181 38ededc0-08b8-5190-f2ac-b31f878777ad