Update the overloaded operator<< for the AddressField enum.
The operator is moved out into the root namespace (where it belongs) and
the implementation is updated to use a lookup table instead of a switch
statement, and to output more information when encountering an invalid
enum value.

AddressFieldTest is added to ensure that the overloaded operator<< works
as expected.

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@191 38ededc0-08b8-5190-f2ac-b31f878777ad