Add address validation functionality to libaddressinput.
This is implemented through the class AddressValidator, which provides
the public interface, and the class ValidationTask which encapsulates
the information necessary to perform validation of one particular
address and call a callback when that has been done.

The interface is asynchronous, allowing it to be used with a Downloader
implementation that waits for several pending requests simultaneously.

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@202 38ededc0-08b8-5190-f2ac-b31f878777ad