Compile only the library with the -fPIC flag.
This patch moves -fPIC flag from target defaults (for all targets) to
the library target, because the unit test executable does not require
it. The move is possible due to the automatic variable "_type", which
contains the value of the "type" configuration for the target.

https://code.google.com/p/gyp/wiki/InputFormatReference#Variables

R=roubert@google.com

Review URL: https://codereview.appspot.com/91470044

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@215 38ededc0-08b8-5190-f2ac-b31f878777ad