Add class PreloadSupplier, implementing the Supplier interface.
PreloadSupplier owns a Retriever object, through which it can load
aggregated address metadata for a region when instructed to, creating
Rule objects and caching these. It also provides methods to check
whether metadata for a particular region is already loaded or in
progress of being loaded.

When using a PreloadSupplier, it becomes possible to do synchronous
address validation using an asynchronous Downloader, and to have full
control over when network access is being done.

R=rouslan@chromium.org
BUG=

Review URL: https://codereview.appspot.com/94700044

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@236 38ededc0-08b8-5190-f2ac-b31f878777ad