Move base tests from MetadataLoaderTest to SupplierTest.
Those tests in MetadataLoaderTest that should hold for all
implementations of the Supplier interface are moved into a new test
fixture Supplier, which is parametrized to be executed with different
wrappers that provide different Supplier implementations.

This verifies that both MetadataLoader and PreloadSupplier implement
this base functionality with the same results.

R=lararennie@google.com, rouslan@chromium.org
BUG=

Review URL: https://codereview.appspot.com/92560044

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@237 38ededc0-08b8-5190-f2ac-b31f878777ad