Parametrize AddressValidatorTest on Supplier to test them all.
The test fixture is parametrized to be executed with different wrappers
that provide different Supplier implementations.

This verifies that both MetadataLoader and PreloadSupplier give the same
results when used for address validation.

In order to do this, add a constructor that takes Supplier* to
AddressValidator.

R=rouslan@chromium.org
BUG=

Review URL: https://codereview.appspot.com/97640045

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@238 38ededc0-08b8-5190-f2ac-b31f878777ad