Refactor the Downloader interface into the Source interface.
Most implementations of this interface don't actually download anything,
so it's more logical to name the interface something else and define it
in terms of lookup keys ("data/CH") instead of URLs, and let those
implementations that do perform network access handle the mapping from
keys to URLs instead.

This simplifies most implementations of this interface and also the
Retriever, which now no longer needs to keep track of the base URL
(which in most cases was bogus anyway).

This change will break all current users of the libaddressinput API.

Ad infinitum et ultra!

TBR=rouslan@chromium.org
BUG=

Review URL: https://codereview.appspot.com/115680043

git-svn-id: http://libaddressinput.googlecode.com/svn/trunk@326 38ededc0-08b8-5190-f2ac-b31f878777ad